+++ src/basics/matrix.hpp
@@ -80,8 +80,8 @@ public:
   inline T& operator[] (const unsigned int j){ 
 #ifdef LISA_DEBUG
     if (j>=m){
-        G_ExceptionList.lthrow("element "+ztos(j)+" of a "+
-                                ztos(m)+"-element vector requested",
+        G_ExceptionList.lthrow("element "+ztos((int)j)+" of a "+
+                                ztos((int)m)+"-element vector requested",
                                 OUT_OF_RANGE);
 	  return contents[0];
 	}
@@ -92,8 +92,8 @@ public:
   inline T operator[](const unsigned int j)const{ 
 #ifdef LISA_DEBUG
     if (j>=m){
-      G_ExceptionList.lthrow("element "+ztos(j)+" of a "+
-				             ztos(m)+"-element vector requested",
+      G_ExceptionList.lthrow("element "+ztos((int)j)+" of a "+
+				             ztos((int)m)+"-element vector requested",
                              OUT_OF_RANGE);
 	  return contents[0];
 	}
@@ -160,8 +160,8 @@ public:
   inline Lisa_Vector<T>& operator[] (const unsigned i){ 
 #ifdef LISA_DEBUG
     if (i>=n){
-        G_ExceptionList.lthrow("row "+ztos(i)+" of a "+
-				               ztos(n)+"-row matrix requested",
+        G_ExceptionList.lthrow("row "+ztos((int)i)+" of a "+
+				               ztos((int)n)+"-row matrix requested",
                                OUT_OF_RANGE);
         return row[0];
 	}
@@ -172,8 +172,8 @@ public:
   inline const Lisa_Vector<T>& operator[](const unsigned i)const{ 
 #ifdef LISA_DEBUG
     if (i>=n){
-	  G_ExceptionList.lthrow("row "+ztos(i)+" of a "+
-				             ztos(n)+"-row matrix requested",
+	  G_ExceptionList.lthrow("row "+ztos((int)i)+" of a "+
+				             ztos((int)n)+"-row matrix requested",
                              OUT_OF_RANGE);
 	  return row[0];
 	}
